+++ Sources/IndexStore/Public/Convenience/IndexStore+ProtocolConvenience.swift
@@ -14,14 +14,15 @@ extension IndexStore {
     /// Will search for protocols that match the given protocol name, then resolve any source symbols for declarations that conform to the given protocol.
     /// - Parameter protocolName: The name of the protocol to search for.
     /// - Returns: Array of ``SourceSymbol``
-    public func sourceSymbols(conformingToProtocol protocolName: String) -> [SourceSymbol] {
+    public func sourceSymbols(conformingToProtocol protocolName: String, recursiveSearchConfig: RecursiveSearchConfiguration = .all) -> [SourceSymbol] {
         // Find any matching protocol symbols
         let query = IndexStoreQuery.protocolDeclarations(matching: protocolName)
             .withIgnoringCase(true)
             .withRestrictingToProjectDirectory(false)
+            .withRecursiveSearchConfig(recursiveSearchConfig)
         let symbols = querySymbols(query)
         // Resolve conforming instances
-        return resolveSymbolsConformingToProtocolSymbols(symbols)
+        return resolveSymbolsConformingToProtocolSymbols(symbols, recursiveSearchConfig: recursiveSearchConfig)
     }
 
     /// Will search for protocols that match the given protocol name, then resolve any source symbols for declarations that conform to the given protocol within the given source files.
@@ -29,20 +30,21 @@ extension IndexStore {
     ///   - protocolName: The name of the protocol to search for.
     ///   - sourceFiles: The source files to search in.
     /// - Returns: Array of ``SourceSymbol``
-    public func sourceSymbols(conformingToProtocol protocolName: String, in sourceFiles: [String]) -> [SourceSymbol] {
+    public func sourceSymbols(conformingToProtocol protocolName: String, in sourceFiles: [String], recursiveSearchConfig: RecursiveSearchConfiguration = .all) -> [SourceSymbol] {
         // Find any matching protocol symbols
         let query = IndexStoreQuery.protocolDeclarations(matching: protocolName)
             .withIgnoringCase(true)
             .withRestrictingToProjectDirectory(false)
+            .withRecursiveSearchConfig(recursiveSearchConfig)
         let symbols = querySymbols(query)
         // Find any matching protocol symbols within the given source files
-        return resolveSymbolsConformingToProtocolSymbols(symbols, in: sourceFiles)
+        return resolveSymbolsConformingToProtocolSymbols(symbols, in: sourceFiles, recursiveSearchConfig: recursiveSearchConfig)
     }
 
     /// Performs the symbol lookups to find symbols conforming to protocols in the given array.
     /// - Parameter symbols: The symbols to search for.
     /// - Returns: Array of ``SourceSymbol`` instances
-    internal func resolveSymbolsConformingToProtocolSymbols(_ symbols: [SourceSymbol]) -> [SourceSymbol] {
+    internal func resolveSymbolsConformingToProtocolSymbols(_ symbols: [SourceSymbol], recursiveSearchConfig: RecursiveSearchConfiguration = .all) -> [SourceSymbol] {
         var results: OrderedSet<SourceSymbol> = []
         symbols.forEach {
             let subclasses = workspace.occurrences(ofUSR: $0.usr, roles: [.baseOf])
@@ -66,7 +68,7 @@ extension IndexStore {
                 guard !results.contains(where: { $0.usr == occurrence.symbol.usr }) else {
                     return
                 }
-                let symbol = sourceSymbolFromOccurrence(occurrence)
+                let symbol = sourceSymbolFromOccurrence(occurrence, recursiveSearchConfig: recursiveSearchConfig)
                 results.append(symbol)
             }
         }
@@ -78,14 +80,18 @@ extension IndexStore {
     ///   - symbols: The symbols to for.
     ///   - sourceFiles: The source files to search in.
     /// - Returns: Array of ``SourceSymbol`` instances
-    internal func resolveSymbolsConformingToProtocolSymbols(_ symbols: [SourceSymbol], in sourceFiles: [String]) -> [SourceSymbol] {
+    internal func resolveSymbolsConformingToProtocolSymbols(
+        _ symbols: [SourceSymbol],
+        in sourceFiles: [String],
+        recursiveSearchConfig: RecursiveSearchConfiguration = .all
+    ) -> [SourceSymbol] {
         var results: OrderedSet<SourceSymbol> = []
         symbols.forEach { symbol in
             let baseOfRelations = workspace.occurrences(ofUSR: symbol.usr, roles: [.baseOf]).flatMap(\.relations)
             let declarationSymbols = workspace.symbolsInSourceFiles(at: sourceFiles, roles: [.definition, .declaration])
             declarationSymbols.forEach { declaration in
                 guard baseOfRelations.contains(where: { $0.symbol.usr == declaration.symbol.usr }) else { return }
-                let sourceSymbol = sourceSymbolFromOccurrence(declaration)
+                let sourceSymbol = sourceSymbolFromOccurrence(declaration, recursiveSearchConfig: recursiveSearchConfig)
                 results.append(sourceSymbol)
             }
         }